+++ xcape/test/fixtures.py
@@ -1,6 +1,8 @@
 import dask
 import dask.array as dsa
 import numpy as np
+import xarray as xr
+import pytest
 
 
 def empty_dask_array(shape, dtype=float, chunks=None):
@@ -13,3 +15,723 @@ def empty_dask_array(shape, dtype=float, chunks=None):
         a = a.rechunk(chunks)
 
     return a
+
+@pytest.fixture(scope='module')
+def dataset_soundings():
+    dtemp ={'attrs': {},
+ 'coords': {'level': {'attrs': {},
+                      'data': [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+                               20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36,
+                               37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53,
+                               54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70,
+                               71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87,
+                               88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103,
+                               104, 105, 106, 107],
+                      'dims': ('level',)}},
+ 'data_vars': {'ML_CAPE_pinc1000_mldepth500': {'attrs': {},
+                                               'data': [10.109490394592285, 3966.124267578125,
+                                                        3043.06494140625, 3.5149593353271484,
+                                                        4004.96875, 4.543837070465088,
+                                                        3091.3564453125, 3.3635528087615967,
+                                                        9.947282791137695, 4129.88916015625,
+                                                        3170.881591796875, 3190.464111328125],
+                                               'dims': ('n2',)},
+               'ML_CIN_pinc1000_mldepth500': {'attrs': {},
+                                              'data': [1.0164976119995117, 160.4442596435547,
+                                                       111.25535583496094, 2.2488975524902344,
+                                                       50.14241409301758, 2.1685924530029297,
+                                                       9.319169998168945, 2.3369221687316895,
+                                                       2.249267339706421, 112.47377014160156,
+                                                       21.510879516601562, 5.772819995880127],
+                                              'dims': ('n2',)},
+               'MU_CAPE_pinc100': {'attrs': {},
+                                   'data': [0.015645720064640045, 4409.52099609375,
+                                            3815.610107421875, 0.0, 4341.2744140625,
+                                            48.773006439208984, 3224.5771484375, 0.0, 0.0,
+                                            5308.6240234375, 4557.6328125, 3913.534912109375],
+                                   'dims': ('n2',)},
+               'MU_CIN_pinc100': {'attrs': {},
+                                  'data': [0.0, 117.04051971435547, 59.330787658691406, 0.0,
+                                           38.938514709472656, 0.0, 0.9654631018638611, 0.0, 0.0,
+                                           49.360198974609375, 0.16132529079914093, 0.0],
+                                  'dims': ('n2',)},
+               'MU_lv_pinc100': {'attrs': {},
+                                 'data': [30.0, 1.0, 1.0, 21.0, 1.0, 1.0, 1.0, 33.0, 30.0, 1.0, 1.0,
+                                          1.0],
+                                 'dims': ('n2',)},
+               'MU_z_pinc100': {'attrs': {},
+                                'data': [16008.916015625, 16082.79296875, 16113.5126953125,
+                                         16325.51171875, 16151.34765625, 16194.9599609375,
+                                         16165.7216796875, 15809.30078125, 15738.236328125,
+                                         16135.2294921875, 16352.34765625, 16324.99609375],
+                                'dims': ('n2',)},
+               'SB_CAPE_pinc100': {'attrs': {},
+                                   'data': [46.99590301513672, 4409.52099609375, 3815.610107421875,
+                                            47.52308654785156, 4341.2744140625, 48.773006439208984,
+                                            3224.5771484375, 49.61775588989258, 48.3080940246582,
+                                            5308.6240234375, 4557.6328125, 3913.534912109375],
+                                   'dims': ('n2',)},
+               'SB_CIN_pinc100': {'attrs': {},
+                                  'data': [1.1928253173828125, 117.04051971435547,
+                                           59.330787658691406, 0.0, 38.938514709472656, 0.0,
+                                           0.9654631018638611, 0.0, 0.0, 49.360198974609375,
+                                           0.16132529079914093, 0.0],
+                                  'dims': ('n2',)},
+               'dewpoint': {'attrs': {},
+                            'data': [[0.0, 20.2, 20.6, 10.0, 20.4, 9.6, 19.8, -3.4, -7.8, 22.4,
+                                      23.6, 22.4],
+                                     [-0.6, 19.8, 20.6, 7.2, 19.2, 8.0, 19.6, -8.0, -10.2, 22.0,
+                                      23.0, 22.2],
+                                     [-3.5, 18.9, 19.1, 6.8, 18.6, 7.6, 19.3, -8.0, -12.0, 19.4,
+                                      21.4, 21.2],
+                                     [-5.6, 18.6, 18.6, 6.2, 18.2, 6.8, 19.2, -8.8, -13.6, 18.6,
+                                      20.2, 20.4],
+                                     [-5.6, 18.4, 17.6, 6.0, 17.6, 6.8, 18.3, -9.8, -13.4, 19.2,
+                                      19.8, 19.8],
+                                     [-5.6, 18.2, 17.4, 5.6, 17.0, 6.8, 17.8, -10.1, -13.6, 18.5,
+                                      18.6, 18.3],
+                                     [-5.6, 18.0, 16.9, 5.2, 16.7, 6.8, 17.7, -10.7, -14.1, 17.3,
+                                      18.4, 17.8],
+                                     [-5.6, 17.3, 16.6, 4.0, 15.2, 6.0, 17.5, -11.2, -14.3, 16.7,
+                                      16.0, 14.4],
+                                     [-5.6, 17.0, 16.3, 4.0, 8.5, 2.2, 16.1, -12.0, -17.5, 16.0,
+                                      14.8, 12.6],
+                                     [-5.9, 16.6, 14.5, 3.4, 8.4, -8.4, 15.4, -15.4, -19.6, 13.6,
+                                      11.6, 12.5],
+                                     [-6.5, 11.4, 13.4, 2.8, 6.8, -8.8, 13.9, -14.9, -19.7, 2.8,
+                                      10.6, 11.7],
+                                     [-7.0, 10.4, 11.7, 2.4, 5.9, -9.2, 10.1, -11.2, -29.1, -4.6,
+                                      9.6, 10.9],
+                                     [-7.4, 8.0, 9.4, 2.3, 2.4, -10.0, 10.0, -10.0, -29.3, -4.8,
+                                      8.8, 10.8],
+                                     [-9.1, 7.4, 6.6, 1.8, 1.7, -8.9, 6.4, -7.8, -34.2, -6.1, 5.6,
+                                      8.0],
+                                     [-11.3, 4.0, 5.4, 1.4, 0.6, -7.6, 3.5, -9.4, -44.1, -7.4, 5.2,
+                                      7.5],
+                                     [-11.3, 2.6, 2.3, 0.9, -0.1, -9.3, 1.4, -11.8, -49.9, -8.7,
+                                      0.8, 6.1],
+                                     [-14.9, 0.0, -0.8, 0.4, -5.8, -17.2, -6.6, -15.1, -49.5, -10.4,
+                                      -2.9, 5.7],
+                                     [-19.3, -0.6, -1.2, -1.7, -3.8, -24.6, -7.8, -15.1, -48.3,
+                                      -12.3, -4.4, 3.2],
+                                     [-32.0, -2.9, -3.4, -2.4, -3.5, -42.8, -9.2, -14.9, -47.5,
+                                      -14.5, -3.4, 4.0],
+                                     [-42.3, -6.4, -9.7, -0.8, -3.4, -38.0, -12.5, -14.9, -46.9,
+                                      -14.6, -4.2, 5.2],
+                                     [-55.3, -9.4, -10.2, -0.6, -8.4, -22.2, -14.2, -16.0, -46.6,
+                                      -14.3, -3.8, 3.0],
+                                     [-55.9, -12.7, -11.8, -2.6, -8.7, -16.4, -14.6, -16.3, -41.5,
+                                      -14.8, 0.4, 2.6],
+                                     [-55.6, -14.3, -13.4, -2.4, -14.6, -18.0, -15.4, -17.8, -38.5,
+                                      -17.8, 0.1, 1.9],
+                                     [-55.5, -14.5, -15.0, -2.4, -14.8, -20.8, -15.4, -18.9, -44.3,
+                                      -19.5, -1.0, 1.8],
+                                     [-57.3, -15.6, -15.4, -5.2, -15.9, -18.1, -17.7, -20.8, -44.7,
+                                      -21.8, 1.2, -0.2],
+                                     [-57.2, -18.1, -18.2, -6.2, -17.5, -10.7, -18.4, -22.1, -38.7,
+                                      -22.5, 0.4, -4.5],
+                                     [-56.9, -18.9, -19.7, -8.7, -17.9, -12.5, -20.6, -22.2, -35.7,
+                                      -21.7, 0.4, -7.1],
+                                     [-57.8, -20.5, -23.1, -10.9, -24.8, -12.9, -24.7, -27.3, -36.8,
+                                      -22.0, -3.0, -8.6],
+                                     [-59.3, -22.1, -24.9, -9.3, -25.3, -16.9, -25.7, -29.4, -38.3,
+                                      -22.7, -4.0, -10.7],
+                                     [-59.5, -22.9, -26.1, -12.3, -28.7, -27.1, -27.5, -33.7, -39.1,
+                                      -25.3, -4.8, -7.8],
+                                     [-60.1, -20.8, -27.5, -12.5, -26.9, -53.1, -28.4, -34.5, -44.1,
+                                      -30.5, -3.6, -7.6],
+                                     [-61.9, -26.8, -30.5, -13.3, -28.5, -53.1, -29.3, -34.1, -44.5,
+                                      -32.9, -5.6, -7.8],
+                                     [-62.4, -28.2, -33.1, -16.5, -26.9, -54.9, -29.7, -36.7, -49.1,
+                                      -33.8, -5.6, -11.1],
+                                     [-64.0, -29.9, -33.3, -16.9, -27.5, -57.3, -35.9, -37.5, -49.9,
+                                      -35.9, -5.6, -12.3],
+                                     [-69.3, -35.5, -34.6, -13.9, -30.1, -57.5, -37.5, -41.4, -49.9,
+                                      -37.1, -2.4, -12.5],
+                                     [-70.5, -36.1, -39.1, -13.9, -31.2, -58.0, -38.5, -50.2, -50.9,
+                                      -42.5, -2.5, -11.7],
+                                     [-73.1, -38.1, -47.5, -16.9, -33.7, -58.3, -38.7, -51.5, -51.5,
+                                      -42.8, -2.9, -13.9],
+                                     [-72.0, -44.0, -50.3, -17.9, -38.1, -49.7, -42.5, -51.7, -56.9,
+                                      -44.3, -7.3, -13.4],
+                                     [-71.5, -47.5, -59.9, -17.9, -47.5, -39.1, -50.1, -51.7, -59.5,
+                                      -44.7, -9.2, -27.1],
+                                     [-71.1, -53.1, -60.3, -19.7, -49.5, -54.7, -53.5, -53.0, -61.2,
+                                      -46.0, -10.2, -26.4],
+                                     [-67.7, -59.7, -69.3, -19.5, -51.9, -57.9, -55.3, -62.2, -67.3,
+                                      -46.3, -12.9, -23.5],
+                                     [-67.3, -61.1, -69.5, -18.1, -54.4, -65.9, -58.1, -62.5, -67.9,
+                                      -46.5, -13.7, -24.9],
+                                     [-67.7, -67.9, -69.7, -20.9, -57.3, -66.4, -60.5, -65.5, -68.0,
+                                      -51.3, -18.9, -27.9],
+                                     [-67.7, -68.5, -72.7, -21.1, -62.9, -66.9, -61.7, -66.2, -68.2,
+                                      -53.9, -22.9, -34.3],
+                                     [-68.7, -65.5, -76.4, -21.4, -65.3, -67.5, -62.5, -69.3, -68.3,
+                                      -55.3, -23.0, -34.5],
+                                     [-68.3, -65.1, -77.1, -22.5, -65.9, -71.6, -67.0, -72.9, -68.3,
+                                      -62.5, -23.3, -36.7],
+                                     [-68.5, -66.1, -76.5, -26.9, -66.5, -72.5, -68.9, -72.7, -68.9,
+                                      -63.5, -29.3, -33.1],
+                                     [-69.1, -66.3, -76.3, -35.5, -68.1, -70.5, -70.3, -72.5, -69.7,
+                                      -64.5, -31.3, -35.3],
+                                     [-69.7, -69.5, -74.7, -35.3, -70.3, -69.8, -72.5, -71.9, -69.9,
+                                      -64.7, -31.4, -30.3],
+                                     [-69.9, -69.1, -74.1, -27.9, -70.5, -66.9, -72.7, -71.5, -70.5,
+                                      -68.1, -31.7, -31.5],
+                                     [-68.9, -67.9, -74.1, -28.3, -71.3, -67.3, -73.1, -71.9, -70.3,
+                                      -67.7, -31.3, -31.6],
+                                     [-69.1, -68.0, -74.0, -28.9, -70.5, -70.5, -70.9, -71.5, -70.9,
+                                      -68.0, -26.9, -33.7],
+                                     [-69.1, -68.9, -73.9, -32.7, -69.9, -70.8, -72.9, -69.9, -71.0,
+                                      -71.1, -28.2, -33.8],
+                                     [-69.3, -68.7, -74.7, -33.7, -70.1, -71.1, -73.1, -68.9, -71.3,
+                                      -70.1, -28.9, -33.9],
+                                     [-69.7, -67.5, -74.9, -35.1, -69.5, -71.1, -73.4, -68.3, -72.3,
+                                      -68.9, -27.6, -30.9],
+                                     [-69.7, -68.4, -73.9, -32.9, -71.1, -72.7, -73.7, -68.9, -72.2,
+                                      -67.5, -27.5, -32.0],
+                                     [-69.7, -69.1, -72.5, -35.5, -71.3, -72.9, -73.7, -67.9, -71.9,
+                                      -66.1, -29.1, -34.9],
+                                     [-69.7, -68.1, -73.3, -38.3, -72.5, -73.1, -73.8, -68.0, -71.3,
+                                      -65.5, -30.8, -36.5],
+                                     [-69.7, -69.5, -74.7, -44.3, -72.1, -72.7, -73.9, -69.5, -71.5,
+                                      -64.5, -34.7, -38.3],
+                                     [-70.7, -69.8, -74.7, -46.5, -72.9, -71.3, -73.5, -68.7, -72.5,
+                                      -64.1, -37.1, -38.5],
+                                     [-71.3, -70.1, -73.5, -53.1, -73.1, -71.5, -77.1, -69.5, -72.5,
+                                      -64.9, -38.8, -40.5],
+                                     [-71.3, -75.3, -75.1, -58.1, -75.5, -71.9, -75.9, -70.5, -72.3,
+                                      -65.1, -43.7, -43.7],
+                                     [-71.5, -74.9, -76.6, -63.7, -74.7, -71.3, -76.5, -70.7, -72.9,
+                                      -64.3, -45.3, -46.3],
+                                     [-71.9, -78.3, -76.9, -65.7, -76.3, -69.5, -76.7, -70.9, -74.1,
+                                      -64.9, -51.3, -52.1],
+                                     [-71.9, -79.7, -75.6, -67.9, -77.5, -70.2, -78.3, -72.3, -74.5,
+                                      -66.3, -53.1, -53.1],
+                                     [-72.9, -79.6, -74.7, -73.7, -77.1, -72.1, -77.3, -73.1, -74.4,
+                                      -67.5, -53.7, -62.7],
+                                     [-72.9, -79.5, -74.5, -73.9, -79.5, -70.9, -76.5, -72.7, -73.5,
+                                      -68.5, -57.3, -69.3],
+                                     [-74.3, -76.8, -74.5, -73.7, -79.4, -71.2, -76.8, -71.5, -74.1,
+                                      -69.4, -63.3, -71.1],
+                                     [-74.8, -76.7, -75.1, -76.1, -79.3, -72.3, -77.6, -71.9, -74.5,
+                                      -72.1, -63.5, -75.7],
+                                     [-75.7, -77.5, -75.0, -75.9, -80.6, -71.9, -77.9, -71.9, -73.9,
+                                      -71.1, -63.7, -76.0],
+                                     [-76.0, -77.0, -74.1, -77.3, -81.7, -72.3, -78.2, -73.5, -74.1,
+                                      -72.5, -65.1, -78.7],
+                                     [-77.1, -76.3, -74.1, -77.7, -81.9, -73.1, -78.3, -73.1, -74.7,
+                                      -72.1, -66.7, -78.9],
+                                     [-78.7, -76.3, -74.0, -77.1, -81.7, -72.7, -77.5, -73.3, -76.1,
+                                      -71.5, -71.1, -79.3],
+                                     [-79.7, -76.3, -73.9, -80.5, -81.5, -74.5, -76.5, -74.6, -76.1,
+                                      -73.7, -71.3, -78.5],
+                                     [-78.3, -76.8, -72.7, -80.3, -77.1, -75.1, -76.1, -75.5, -76.0,
+                                      -73.9, -75.3, -78.9],
+                                     [-77.7, -78.1, -70.1, -80.1, -77.7, -75.3, -73.9, -75.0, -75.9,
+                                      -74.0, -75.3, -79.9],
+                                     [-77.6, -77.1, -70.7, -77.1, -77.7, -76.7, -73.5, -74.7, -77.3,
+                                      -74.2, -79.3, -80.4],
+                                     [-76.5, -75.5, -70.7, -77.0, -75.9, -76.7, -71.7, -72.6, -76.7,
+                                      -74.3, -81.4, -80.7],
+                                     [-79.3, -75.1, -70.9, -76.7, -75.5, -77.1, -71.7, -71.9, -76.5,
+                                      -73.5, -81.9, -78.7],
+                                     [-78.5, -74.3, -69.1, -76.1, -73.1, -77.9, -72.2, -72.1, -77.7,
+                                      -72.1, -82.5, -78.9],
+                                     [-77.1, -73.7, -69.1, -74.7, -72.5, -79.7, -72.7, -72.3, -77.9,
+                                      -71.7, -82.7, -79.9],
+                                     [-77.1, -73.4, -69.1, -73.7, -72.5, -80.3, -72.7, -73.5, -78.1,
+                                      -71.9, -81.1, -78.3],
+                                     [-77.1, -72.7, -69.1, -74.1, -74.3, -80.5, -73.1, -78.5, -77.3,
+                                      -72.1, -83.7, -79.5],
+                                     [-76.9, -73.7, -68.5, -74.1, -74.9, -78.9, -72.8, -80.1, -77.9,
+                                      -73.9, -84.3, -81.3],
+                                     [-76.1, -75.6, -69.5, -74.0, -75.9, -79.9, -72.6, -77.6, -77.9,
+                                      -73.6, -84.0, -77.3],
+                                     [-76.9, -76.3, -68.1, -72.1, -73.9, -79.8, -72.3, -77.5, -79.9,
+                                      -72.3, -83.9, -78.9],
+                                     [-77.1, -76.6, -69.3, -72.0, -73.8, -79.3, -72.0, -77.3, -79.8,
+                                      -72.3, -82.8, -73.5],
+                                     [-78.2, -78.7, -68.7, -71.5, -73.7, -79.8, -71.7, -75.7, -79.3,
+                                      -73.1, -82.3, -72.8],
+                                     [-79.1, -84.3, -67.5, -71.2, -73.7, -80.6, -71.3, -73.8, -79.3,
+                                      -73.0, -80.6, -72.7],
+                                     [-78.7, -87.3, -67.6, -70.8, -71.8, -80.9, -70.9, -73.3, -79.3,
+                                      -72.1, -80.2, -72.1],
+                                     [-75.9, -87.7, -68.5, -70.5, -71.1, -81.5, -71.1, -73.9, -77.8,
+                                      -71.8, -79.5, -72.9],
+                                     [-75.9, -87.7, -68.7, -70.5, -72.0, -81.0, -71.3, -74.5, -76.5,
+                                      -70.9, -78.9, -71.5],
+                                     [-76.9, -87.6, -67.3, -70.5, -72.6, -80.9, -70.9, -74.4, -76.1,
+                                      -71.7, -78.8, -71.3],
+                                     [-76.8, -87.3, -67.6, -69.4, -72.9, -79.1, -71.1, -74.1, -76.1,
+                                      -71.9, -77.7, -71.8],
+                                     [-75.7, -85.5, -69.3, -67.9, -70.7, -78.1, -71.1, -73.8, -76.7,
+                                      -71.8, -77.4, -72.4],
+                                     [-76.1, -85.4, -69.3, -67.9, -69.5, -79.7, -73.5, -71.9, -76.9,
+                                      -71.2, -77.2, -72.9],
+                                     [-76.1, -85.3, -69.5, -67.9, -70.0, -79.9, -73.0, -73.3, -77.6,
+                                      -70.6, -76.7, -71.3],
+                                     [-75.5, -81.9, -69.8, -67.9, -70.5, -80.8, -72.9, -74.3, -78.3,
+                                      -70.0, -76.0, -71.5],
+                                     [-74.3, -81.9, -70.3, -66.1, -72.7, -81.3, -72.3, -74.0, -78.7,
+                                      -69.7, -75.9, -71.9],
+                                     [-76.1, -81.9, -70.6, -65.9, -75.5, -81.3, -74.7, -72.9, -78.6,
+                                      -71.1, -75.3, -71.9],
+                                     [-75.5, -81.9, -70.7, -65.9, -84.2, -82.0, -74.9, -71.3, -78.5,
+                                      -70.3, -75.0, -71.9],
+                                     [-75.5, -81.4, -68.1, -65.9, -87.1, -82.1, -75.2, -72.7, -78.7,
+                                      -70.6, -74.6, -72.5],
+                                     [-75.7, -80.2, -69.4, -65.9, -85.1, -86.7, -76.2, -71.5, -78.5,
+                                      -71.8, -74.2, -73.9],
+                                     [-77.1, -79.0, -70.9, -65.9, -83.8, -87.1, -76.7, -67.3, -78.3,
+                                      -73.1, -74.0, -74.5],
+                                     [-78.4, -77.7, -71.3, -65.9, -83.2, -87.8, -76.8, -68.1, -79.1,
+                                      -73.3, -73.7, -73.5],
+                                     [-84.3, -76.5, -70.3, -66.3, -80.9, -89.7, -77.7, -68.5, -79.3,
+                                      -73.4, -73.3, -73.6],
+                                     [-86.7, -76.7, -69.1, -66.7, -80.1, -89.9, -78.7, -66.5, -80.1,
+                                      -73.5, -73.1, -74.4],
+                                     [-86.7, -77.0, -68.3, -67.7, -81.0, -88.2, -71.1, -66.5, -80.4,
+                                      -74.2, -72.6, -74.5]],
+                            'dims': ('level', 'n2')},
+               'pressure': {'attrs': {},
+                            'data': [[984.0, 962.0, 962.0, 967.0, 969.0, 979.0, 961.0, 974.0, 979.0,
+                                      966.0, 974.0, 973.0],
+                                     [980.0, 959.0, 957.0, 947.0, 940.5, 962.0, 953.0, 970.0, 962.0,
+                                      963.0, 972.0, 968.0],
+                                     [953.8, 933.6, 933.5, 939.0, 925.0, 949.5, 932.8, 957.0, 947.4,
+                                      937.8, 945.5, 944.8],
+                                     [936.0, 925.0, 925.0, 925.0, 908.6, 925.0, 925.0, 943.6, 934.0,
+                                      930.0, 925.0, 925.0],
+                                     [925.0, 901.8, 914.0, 907.6, 877.6, 916.6, 901.3, 925.0, 925.0,
+                                      925.0, 913.8, 913.2],
+                                     [920.0, 870.9, 901.8, 877.1, 852.0, 884.4, 890.0, 909.5, 912.0,
+                                      906.0, 886.0, 882.3],
+                                     [887.1, 850.0, 870.9, 850.0, 850.0, 850.0, 870.5, 876.4, 877.6,
+                                      875.0, 882.9, 872.0],
+                                     [850.0, 826.0, 850.0, 818.6, 833.0, 841.0, 850.0, 850.0, 860.0,
+                                      861.0, 850.0, 850.0],
+                                     [839.0, 820.0, 843.0, 818.0, 818.3, 833.0, 811.4, 841.0, 850.0,
+                                      855.0, 848.0, 828.0],
+                                     [824.1, 819.0, 811.7, 790.3, 818.0, 830.0, 794.0, 817.0, 844.4,
+                                      850.0, 823.5, 822.7],
+                                     [793.8, 816.0, 793.0, 763.0, 793.0, 826.0, 783.0, 813.3, 844.0,
+                                      837.0, 816.0, 794.0],
+                                     [771.0, 815.0, 783.5, 746.0, 789.8, 823.1, 756.0, 786.0, 838.0,
+                                      821.0, 794.9, 767.0],
+                                     [764.5, 811.8, 771.0, 736.3, 777.0, 816.0, 755.5, 783.2, 828.0,
+                                      815.8, 778.0, 766.3],
+                                     [735.7, 811.0, 767.0, 700.0, 762.2, 793.7, 741.0, 778.0, 812.1,
+                                      787.1, 767.3, 739.2],
+                                     [700.0, 799.0, 756.2, 685.0, 739.0, 768.0, 728.7, 771.0, 780.9,
+                                      759.5, 766.0, 734.0],
+                                     [690.0, 783.6, 729.4, 660.4, 735.3, 765.3, 720.0, 754.2, 763.0,
+                                      732.9, 758.0, 702.0],
+                                     [670.0, 756.4, 703.6, 642.0, 708.0, 753.0, 717.0, 731.0, 750.8,
+                                      700.0, 740.3, 700.0],
+                                     [661.0, 750.0, 700.0, 613.4, 700.0, 737.6, 713.0, 726.0, 721.5,
+                                      656.6, 733.0, 689.0],
+                                     [654.9, 729.8, 680.0, 604.0, 683.6, 701.0, 700.0, 700.0, 700.0,
+                                      609.5, 716.0, 687.1],
+                                     [650.0, 700.0, 653.9, 590.8, 676.0, 700.0, 653.0, 698.7, 696.0,
+                                      608.0, 700.0, 684.0],
+                                     [642.0, 653.7, 652.0, 589.0, 661.0, 696.0, 629.4, 672.2, 693.2,
+                                      571.0, 676.0, 662.2],
+                                     [623.0, 607.0, 637.0, 575.0, 658.8, 673.0, 624.0, 665.0, 638.8,
+                                      564.5, 669.0, 658.0],
+                                     [605.6, 584.8, 629.0, 568.9, 611.4, 659.5, 606.4, 646.5, 609.0,
+                                      521.9, 664.1, 638.1],
+                                     [601.0, 582.0, 607.1, 568.0, 610.0, 637.0, 606.0, 633.0, 588.3,
+                                      500.0, 649.0, 634.0],
+                                     [561.0, 562.7, 601.0, 556.0, 565.9, 634.0, 584.0, 597.2, 587.0,
+                                      482.2, 640.0, 614.5],
+                                     [559.8, 520.0, 583.0, 547.5, 504.0, 630.0, 578.0, 574.0, 573.0,
+                                      477.0, 633.0, 576.0],
+                                     [553.0, 500.0, 562.8, 526.0, 500.0, 612.0, 562.3, 573.8, 559.0,
+                                      469.0, 628.0, 569.7],
+                                     [537.9, 480.8, 521.4, 523.0, 484.2, 611.3, 533.0, 562.0, 541.1,
+                                      463.2, 623.0, 566.0],
+                                     [515.0, 461.8, 500.0, 519.0, 483.0, 605.0, 520.9, 551.0, 519.0,
+                                      452.0, 616.8, 539.0],
+                                     [504.0, 453.0, 482.3, 515.0, 452.0, 571.0, 500.0, 529.0, 500.0,
+                                      444.0, 612.0, 533.0],
+                                     [500.0, 445.0, 463.6, 500.0, 442.0, 566.0, 481.9, 519.0, 459.0,
+                                      417.0, 603.0, 529.0],
+                                     [483.0, 408.5, 425.0, 493.0, 430.0, 565.9, 463.2, 509.0, 456.7,
+                                      400.0, 597.0, 527.7],
+                                     [476.7, 400.0, 400.0, 487.7, 414.0, 538.0, 455.0, 500.0, 433.0,
+                                      392.9, 594.2, 500.0],
+                                     [457.2, 391.8, 398.0, 487.0, 411.3, 503.2, 410.5, 494.0, 429.0,
+                                      376.6, 582.0, 488.2],
+                                     [400.0, 365.0, 394.6, 478.0, 400.0, 500.0, 400.0, 467.1, 424.0,
+                                      368.0, 574.0, 486.0],
+                                     [386.7, 359.9, 383.0, 475.0, 394.5, 483.4, 394.1, 411.8, 421.0,
+                                      350.0, 572.2, 482.0],
+                                     [360.0, 344.0, 318.9, 472.0, 382.0, 471.0, 393.0, 404.0, 418.8,
+                                      345.8, 561.0, 474.0],
+                                     [324.3, 315.8, 300.0, 469.0, 371.0, 454.0, 346.0, 400.0, 400.0,
+                                      321.0, 529.0, 470.0],
+                                     [310.1, 300.0, 253.0, 469.0, 318.1, 437.0, 319.2, 390.0, 380.0,
+                                      316.8, 500.0, 453.0],
+                                     [300.0, 275.8, 250.0, 452.0, 308.0, 410.3, 308.0, 377.7, 367.4,
+                                      303.1, 490.7, 451.4],
+                                     [287.0, 250.0, 203.0, 450.9, 300.0, 405.0, 300.0, 302.5, 325.0,
+                                      300.0, 465.0, 445.0],
+                                     [282.0, 245.0, 201.5, 443.0, 291.1, 400.0, 272.0, 300.0, 300.0,
+                                      289.0, 458.0, 439.0],
+                                     [270.3, 211.0, 200.0, 435.0, 281.0, 393.4, 253.0, 282.0, 298.0,
+                                      265.0, 443.0, 419.0],
+                                     [270.0, 207.0, 179.0, 420.0, 250.0, 387.0, 250.0, 275.8, 293.6,
+                                      250.0, 440.0, 400.5],
+                                     [255.0, 206.0, 165.4, 416.4, 231.0, 377.0, 243.0, 250.0, 291.0,
+                                      240.0, 436.1, 400.0],
+                                     [250.0, 204.0, 163.0, 405.0, 220.4, 317.1, 211.6, 227.0, 285.0,
+                                      209.8, 421.0, 397.0],
+                                     [237.0, 200.0, 150.0, 400.0, 209.0, 305.0, 200.0, 221.0, 272.0,
+                                      206.0, 413.0, 394.0],
+                                     [234.0, 197.7, 144.0, 387.0, 200.0, 300.0, 189.0, 220.0, 250.0,
+                                      200.0, 404.0, 391.0],
+                                     [224.3, 171.0, 142.0, 381.0, 182.0, 290.1, 180.0, 216.9, 249.0,
+                                      197.0, 402.6, 389.0],
+                                     [220.0, 166.0, 136.0, 377.0, 179.0, 252.0, 174.4, 215.0, 235.0,
+                                      185.0, 400.0, 386.0],
+                                     [218.0, 164.0, 135.1, 371.0, 166.0, 250.0, 166.0, 212.0, 227.0,
+                                      183.0, 396.0, 384.5],
+                                     [214.0, 162.7, 128.5, 368.6, 161.0, 224.0, 158.0, 211.0, 218.0,
+                                      181.2, 392.0, 358.0],
+                                     [213.0, 151.0, 126.0, 353.7, 159.0, 219.4, 150.0, 209.0, 212.2,
+                                      166.0, 386.7, 354.1],
+                                     [210.0, 150.0, 122.2, 350.0, 152.0, 214.0, 149.0, 208.0, 200.0,
+                                      165.0, 384.0, 346.0],
+                                     [205.0, 146.0, 121.0, 336.0, 150.0, 209.0, 143.2, 200.0, 194.0,
+                                      164.1, 371.2, 334.0],
+                                     [201.0, 140.6, 116.0, 331.0, 141.0, 200.0, 138.0, 192.0, 193.0,
+                                      163.0, 370.0, 325.5],
+                                     [200.0, 136.0, 114.0, 325.2, 135.0, 199.0, 126.0, 189.0, 190.0,
+                                      157.0, 363.0, 304.0],
+                                     [197.0, 132.0, 108.0, 319.0, 132.0, 198.0, 123.4, 187.8, 183.0,
+                                      156.2, 356.3, 300.0],
+                                     [194.0, 125.0, 106.0, 300.0, 130.0, 194.0, 121.0, 175.0, 171.0,
+                                      155.0, 341.8, 286.4],
+                                     [183.0, 121.4, 100.0, 292.0, 124.0, 191.0, 118.0, 165.0, 161.0,
+                                      153.0, 333.0, 285.0],
+                                     [166.0, 117.0, 95.9, 250.0, 122.9, 189.6, 108.0, 155.0, 159.9,
+                                      150.0, 327.8, 274.1],
+                                     [158.0, 106.0, 86.1, 226.0, 111.0, 187.0, 105.0, 151.0, 157.0,
+                                      149.0, 313.0, 258.0],
+                                     [155.0, 100.0, 77.9, 200.0, 105.0, 183.0, 100.0, 150.0, 150.0,
+                                      147.0, 300.0, 250.0],
+                                     [150.0, 89.9, 76.5, 190.0, 100.0, 177.0, 89.6, 145.0, 143.0,
+                                      141.0, 264.2, 229.4],
+                                     [148.0, 85.9, 74.0, 177.0, 94.7, 172.2, 86.3, 137.0, 140.0,
+                                      132.0, 254.0, 226.0],
+                                     [139.4, 81.2, 72.5, 153.0, 87.6, 161.0, 82.6, 130.0, 138.7,
+                                      128.7, 250.0, 200.0],
+                                     [139.0, 78.9, 70.4, 150.0, 82.7, 158.0, 79.8, 127.7, 131.0,
+                                      126.0, 230.8, 172.8],
+                                     [129.0, 77.2, 70.0, 148.0, 78.1, 156.6, 78.5, 121.0, 126.1,
+                                      122.5, 202.0, 166.0],
+                                     [126.4, 77.1, 67.2, 137.0, 75.9, 152.0, 74.6, 116.0, 123.0,
+                                      113.0, 201.0, 150.0],
+                                     [122.0, 75.1, 67.0, 132.0, 74.2, 150.0, 73.4, 115.9, 120.0,
+                                      111.0, 200.0, 148.9],
+                                     [120.3, 73.3, 65.0, 127.4, 72.7, 148.0, 70.9, 110.0, 117.0,
+                                      104.0, 191.7, 137.0],
+                                     [114.4, 70.7, 63.7, 126.0, 70.0, 142.0, 70.0, 108.0, 114.5,
+                                      100.0, 183.0, 134.5],
+                                     [106.0, 70.0, 57.7, 124.0, 66.9, 140.0, 67.4, 100.0, 109.1,
+                                      95.3, 165.8, 130.0],
+                                     [102.0, 69.7, 54.4, 107.0, 65.1, 128.0, 64.4, 90.7, 109.0,
+                                      91.0, 165.0, 129.0],
+                                     [100.0, 66.3, 52.4, 105.0, 62.5, 125.0, 64.1, 85.1, 107.0,
+                                      90.7, 150.1, 127.7],
+                                     [98.4, 58.6, 50.9, 100.0, 60.4, 123.3, 62.2, 83.1, 105.0, 86.5,
+                                      150.0, 125.0],
+                                     [98.0, 56.9, 50.0, 89.6, 60.4, 114.0, 61.0, 82.2, 100.0, 78.3,
+                                      135.5, 121.3],
+                                     [95.1, 54.1, 49.7, 88.8, 57.4, 111.0, 55.3, 74.5, 89.7, 75.1,
+                                      128.7, 119.0],
+                                     [81.7, 53.5, 47.3, 84.4, 56.7, 106.0, 55.2, 72.2, 86.9, 74.5,
+                                      127.0, 116.0],
+                                     [80.5, 51.5, 45.9, 78.4, 54.2, 100.0, 52.6, 70.9, 81.3, 73.3,
+                                      122.1, 115.2],
+                                     [78.4, 50.0, 45.1, 76.3, 52.0, 94.0, 50.1, 70.0, 80.3, 71.9,
+                                      121.0, 111.0],
+                                     [76.5, 49.0, 43.0, 74.8, 51.9, 87.3, 50.0, 67.5, 79.6, 70.9,
+                                      118.0, 107.0],
+                                     [73.2, 46.8, 42.3, 70.0, 50.0, 85.1, 48.6, 57.7, 72.0, 70.0,
+                                      107.0, 100.0],
+                                     [72.8, 44.4, 40.8, 69.3, 49.5, 82.5, 45.4, 51.5, 70.1, 64.8,
+                                      103.0, 94.2],
+                                     [71.5, 40.3, 38.6, 69.0, 48.5, 79.6, 43.3, 50.1, 70.0, 64.2,
+                                      101.0, 88.3],
+                                     [70.0, 38.9, 34.0, 63.6, 47.3, 79.0, 40.8, 50.0, 60.7, 61.2,
+                                      100.0, 82.3],
+                                     [69.3, 38.3, 32.6, 62.5, 44.8, 77.3, 39.3, 49.6, 60.5, 61.0,
+                                      92.7, 78.9],
+                                     [65.9, 34.5, 30.0, 59.4, 42.6, 75.2, 37.4, 49.0, 58.9, 58.2,
+                                      89.2, 76.5],
+                                     [63.1, 31.2, 28.2, 56.6, 42.4, 71.5, 35.7, 47.7, 54.8, 58.1,
+                                      79.1, 76.0],
+                                     [62.6, 30.0, 27.9, 53.8, 40.6, 70.0, 33.6, 47.3, 54.6, 56.4,
+                                      76.7, 70.0],
+                                     [59.7, 27.7, 24.2, 52.2, 40.0, 67.3, 32.4, 45.4, 52.1, 55.3,
+                                      73.0, 65.9],
+                                     [59.6, 27.3, 23.4, 51.3, 36.8, 64.6, 30.9, 43.7, 50.0, 51.7,
+                                      70.0, 62.7],
+                                     [54.3, 26.1, 22.4, 50.0, 35.1, 64.2, 30.0, 43.2, 49.7, 50.0,
+                                      69.6, 62.2],
+                                     [53.9, 22.1, 22.0, 48.8, 34.2, 62.1, 28.1, 41.4, 49.7, 48.3,
+                                      64.3, 59.7],
+                                     [50.0, 20.0, 20.0, 47.2, 31.8, 59.5, 28.1, 41.1, 48.0, 47.7,
+                                      62.9, 56.9],
+                                     [49.0, 19.6, 20.0, 46.5, 30.7, 51.7, 23.7, 39.4, 47.4, 45.4,
+                                      61.9, 54.6],
+                                     [44.9, 19.3, 18.2, 44.3, 30.4, 50.0, 22.2, 37.3, 45.1, 43.3,
+                                      59.9, 52.3],
+                                     [44.3, 18.0, 17.4, 43.7, 30.0, 45.7, 22.0, 36.0, 43.0, 41.2,
+                                      57.2, 51.6],
+                                     [43.2, 17.0, 15.9, 40.3, 28.9, 43.5, 21.4, 35.6, 41.7, 40.3,
+                                      56.6, 50.0],
+                                     [39.3, 15.5, 15.2, 39.8, 27.6, 43.5, 20.0, 34.2, 41.0, 37.0,
+                                      54.3, 49.1],
+                                     [35.6, 15.1, 15.0, 38.4, 23.9, 41.5, 19.9, 33.2, 39.7, 36.0,
+                                      53.2, 47.9],
+                                     [31.6, 14.8, 14.3, 36.6, 22.8, 41.2, 19.3, 31.2, 38.0, 35.6,
+                                      51.7, 46.8],
+                                     [30.0, 14.2, 12.6, 33.3, 20.0, 37.6, 17.6, 30.0, 37.2, 34.0,
+                                      50.0, 44.6],
+                                     [27.8, 13.6, 11.0, 31.8, 18.1, 37.3, 16.9, 28.9, 36.2, 32.4,
+                                      49.3, 43.6],
+                                     [27.3, 13.0, 10.6, 30.0, 17.2, 35.8, 16.8, 28.0, 35.8, 32.1,
+                                      48.3, 42.6],
+                                     [25.0, 12.4, 10.0, 29.4, 14.3, 32.4, 16.1, 27.6, 35.4, 30.8,
+                                      47.0, 42.5],
+                                     [22.6, 11.8, 9.6, 28.9, 13.5, 31.4, 15.2, 26.8, 34.3, 30.0,
+                                      46.5, 40.5],
+                                     [22.6, 11.3, 9.4, 27.6, 13.1, 31.0, 13.1, 26.7, 33.8, 29.4,
+                                      44.8, 40.2]],
+                            'dims': ('level', 'n2')},
+               'temperature': {'attrs': {},
+                               'data': [[21.0, 30.2, 30.6, 39.0, 31.4, 26.6, 31.8, 16.6, 6.2, 29.4,
+                                         33.6, 33.4],
+                                        [20.4, 29.8, 29.6, 36.2, 29.0, 24.0, 30.6, 16.0, 3.8, 29.0,
+                                         33.0, 32.2],
+                                        [18.0, 27.4, 27.4, 35.5, 27.6, 22.9, 28.9, 14.0, 2.6, 27.2,
+                                         30.9, 30.2],
+                                        [16.4, 26.6, 26.6, 34.2, 26.2, 20.8, 28.2, 12.8, 1.4, 26.6,
+                                         29.2, 28.4],
+                                        [15.4, 24.6, 25.6, 32.6, 23.4, 20.0, 25.9, 11.2, 0.6, 26.2,
+                                         28.2, 27.4],
+                                        [14.9, 21.8, 24.6, 29.8, 21.0, 17.1, 24.8, 9.9, -0.5, 24.6,
+                                         25.6, 24.7],
+                                        [11.9, 19.8, 22.1, 27.2, 20.8, 13.8, 23.2, 7.1, -3.6, 22.0,
+                                         25.4, 23.8],
+                                        [8.4, 17.8, 20.4, 24.1, 19.6, 13.0, 21.4, 4.8, -5.3, 20.8,
+                                         23.0, 22.4],
+                                        [7.4, 17.8, 19.8, 24.0, 20.4, 13.2, 17.7, 4.0, -5.5, 21.0,
+                                         22.8, 20.6],
+                                        [6.0, 18.4, 19.6, 21.9, 20.4, 13.6, 16.0, 3.6, -5.7, 21.6,
+                                         21.9, 20.1],
+                                        [3.2, 19.4, 19.4, 19.8, 18.8, 14.2, 15.3, 3.4, -5.7, 23.8,
+                                         21.6, 17.6],
+                                        [1.0, 20.4, 19.0, 18.4, 18.7, 14.1, 13.4, 1.8, -5.1, 23.4,
+                                         19.5, 15.2],
+                                        [0.5, 21.2, 18.4, 17.5, 18.4, 14.0, 13.4, 1.9, -5.3, 22.9,
+                                         17.8, 15.1],
+                                        [-2.1, 21.4, 18.6, 13.8, 16.9, 12.3, 14.4, 2.2, -6.2, 20.3,
+                                         17.3, 12.8],
+                                        [-5.3, 22.0, 17.5, 12.2, 14.6, 10.4, 14.4, 2.6, -7.9, 17.6,
+                                         17.2, 12.4],
+                                        [-6.3, 20.9, 14.9, 9.5, 14.3, 10.3, 14.4, 1.1, -8.9, 15.0,
+                                         17.8, 9.6],
+                                        [-7.9, 18.9, 12.2, 7.4, 12.2, 9.8, 14.4, -1.1, -9.8, 11.6,
+                                         17.0, 9.6],
+                                        [-8.3, 18.4, 11.8, 4.6, 11.2, 8.8, 14.2, -1.4, -11.9, 7.0,
+                                         16.6, 9.2],
+                                        [-8.3, 16.5, 9.6, 3.6, 9.4, 6.2, 12.8, -2.9, -13.5, 1.6,
+                                         15.6, 9.1],
+                                        [-8.3, 13.6, 8.8, 2.4, 8.6, 6.0, 8.3, -3.0, -13.9, 1.4,
+                                         13.8, 8.8],
+                                        [-7.3, 8.4, 8.8, 2.2, 7.6, 5.8, 6.0, -4.8, -14.2, -3.3,
+                                         11.2, 7.3],
+                                        [-8.9, 2.7, 7.2, 0.8, 7.4, 3.6, 5.4, -5.3, -19.4, -4.0,
+                                         10.4, 7.0],
+                                        [-7.8, -0.1, 7.6, 0.0, 2.4, 2.4, 4.6, -7.4, -22.5, -8.8,
+                                         9.8, 5.0],
+                                        [-7.5, -0.5, 5.3, -0.1, 2.2, 0.2, 4.6, -8.9, -23.6, -11.5,
+                                         8.0, 4.6],
+                                        [-10.3, -3.1, 4.6, -0.9, -3.2, -0.1, 2.3, -13.2, -23.7,
+                                         -13.8, 7.2, 2.9],
+                                        [-10.2, -9.1, 3.8, -1.7, -11.5, -0.7, 1.6, -16.1, -22.7,
+                                         -14.5, 6.4, -0.7],
+                                        [-9.9, -11.9, 1.6, -3.7, -11.9, -2.5, 0.5, -16.1, -23.7,
+                                         -15.7, 6.4, -0.0],
+                                        [-11.6, -14.6, -3.2, -3.9, -13.2, -2.5, -1.7, -16.3, -24.4,
+                                         -16.4, 6.0, 0.4],
+                                        [-14.3, -17.4, -5.9, -4.3, -13.3, -2.9, -3.1, -17.4, -25.3,
+                                         -17.9, 5.5, -2.7],
+                                        [-14.5, -18.7, -8.2, -4.3, -17.7, -5.1, -5.5, -19.7, -27.1,
+                                         -18.3, 5.2, -3.3],
+                                        [-15.1, -17.5, -10.8, -5.5, -18.9, -5.1, -7.5, -20.5, -31.1,
+                                         -21.5, 4.4, -3.7],
+                                        [-16.9, -22.6, -16.5, -6.3, -20.5, -5.1, -9.7, -19.1, -31.3,
+                                         -23.9, 4.4, -3.8],
+                                        [-17.7, -23.9, -20.1, -6.8, -21.9, -6.9, -10.7, -19.7,
+                                         -33.1, -25.0, 4.0, -6.7],
+                                        [-20.2, -25.2, -20.3, -6.9, -22.1, -11.1, -17.7, -19.5,
+                                         -32.9, -27.7, 2.4, -7.9],
+                                        [-28.3, -29.5, -20.7, -7.9, -23.1, -11.5, -19.5, -22.6,
+                                         -30.9, -29.1, 1.6, -8.1],
+                                        [-30.5, -30.1, -22.1, -7.9, -23.9, -13.7, -20.5, -29.4,
+                                         -30.9, -31.5, 1.4, -8.3],
+                                        [-35.1, -32.1, -33.5, -7.9, -25.7, -15.3, -20.7, -30.5,
+                                         -31.2, -32.2, 0.0, -8.9],
+                                        [-40.8, -37.4, -37.3, -7.9, -27.1, -17.7, -28.5, -30.7,
+                                         -33.9, -36.3, -2.7, -9.1],
+                                        [-43.3, -40.5, -47.9, -7.9, -36.5, -20.1, -33.4, -30.7,
+                                         -36.5, -36.9, -6.1, -10.1],
+                                        [-45.1, -45.7, -48.3, -9.7, -38.5, -23.3, -35.5, -32.6,
+                                         -37.5, -38.8, -7.0, -10.2],
+                                        [-47.7, -51.7, -59.3, -9.9, -39.9, -23.9, -37.3, -46.0,
+                                         -41.3, -39.3, -9.7, -10.5],
+                                        [-48.3, -53.1, -59.5, -11.1, -41.5, -23.9, -43.1, -46.5,
+                                         -44.9, -40.5, -8.7, -10.9],
+                                        [-47.7, -60.9, -59.7, -11.9, -43.3, -24.4, -46.5, -50.5,
+                                         -45.2, -45.3, -9.9, -13.9],
+                                        [-47.7, -61.5, -63.7, -14.1, -49.9, -24.9, -46.7, -51.6,
+                                         -45.9, -47.9, -9.9, -16.4],
+                                        [-48.7, -58.5, -67.4, -14.4, -54.3, -26.5, -47.5, -56.3,
+                                         -46.3, -49.3, -10.4, -16.5],
+                                        [-49.3, -58.1, -68.1, -15.5, -54.9, -37.1, -54.9, -60.9,
+                                         -46.3, -56.5, -12.3, -16.7],
+                                        [-51.5, -58.1, -67.5, -15.9, -55.5, -39.5, -57.9, -60.7,
+                                         -47.9, -57.5, -13.3, -17.1],
+                                        [-51.1, -58.4, -67.3, -17.5, -57.1, -40.5, -59.3, -60.5,
+                                         -48.7, -58.5, -14.3, -17.3],
+                                        [-52.3, -61.5, -65.7, -18.3, -60.3, -42.5, -61.5, -59.9,
+                                         -48.9, -58.7, -14.4, -17.3],
+                                        [-52.9, -61.1, -65.1, -18.9, -59.5, -50.9, -62.1, -59.5,
+                                         -51.5, -62.1, -14.7, -17.5],
+                                        [-51.9, -59.9, -65.1, -19.3, -61.3, -51.3, -63.1, -59.9,
+                                         -50.3, -61.7, -15.3, -17.8],
+                                        [-52.1, -59.9, -65.0, -19.7, -60.5, -56.5, -59.9, -59.5,
+                                         -51.9, -62.0, -15.9, -22.7],
+                                        [-52.1, -59.9, -64.9, -22.1, -58.9, -57.2, -61.9, -56.9,
+                                         -52.3, -65.1, -16.6, -23.1],
+                                        [-52.7, -59.7, -65.7, -22.7, -59.1, -58.1, -62.1, -55.9,
+                                         -53.3, -64.1, -16.9, -23.9],
+                                        [-53.7, -58.5, -65.9, -25.1, -58.5, -58.1, -62.9, -55.3,
+                                         -54.3, -62.9, -18.4, -24.9],
+                                        [-52.7, -59.4, -64.9, -25.9, -60.1, -59.7, -63.7, -54.9,
+                                         -54.2, -61.5, -18.5, -26.3],
+                                        [-52.7, -60.1, -63.5, -26.6, -60.3, -59.9, -62.7, -53.9,
+                                         -53.9, -60.1, -20.1, -29.9],
+                                        [-52.7, -58.1, -64.3, -27.3, -61.5, -60.1, -63.3, -54.1,
+                                         -50.3, -59.5, -21.0, -30.5],
+                                        [-51.7, -59.5, -65.7, -30.3, -61.1, -59.7, -63.9, -56.5,
+                                         -49.5, -58.5, -22.9, -32.3],
+                                        [-52.7, -59.8, -65.7, -31.5, -61.9, -57.3, -63.5, -54.7,
+                                         -51.5, -57.1, -24.1, -32.5],
+                                        [-56.3, -60.1, -64.5, -41.1, -62.2, -57.5, -68.1, -55.5,
+                                         -51.5, -57.9, -25.0, -34.9],
+                                        [-56.3, -65.3, -66.1, -47.1, -65.5, -57.9, -65.9, -56.5,
+                                         -51.3, -58.1, -27.7, -38.7],
+                                        [-55.5, -64.9, -67.6, -53.7, -64.7, -57.3, -66.5, -56.7,
+                                         -53.9, -57.3, -30.3, -40.3],
+                                        [-55.9, -69.7, -67.9, -56.7, -66.3, -54.5, -66.7, -56.9,
+                                         -56.1, -56.9, -37.8, -45.2],
+                                        [-55.9, -71.7, -66.6, -58.9, -68.5, -55.0, -68.3, -59.3,
+                                         -56.5, -58.3, -40.1, -46.1],
+                                        [-57.8, -70.9, -65.7, -65.7, -68.1, -56.1, -67.3, -60.1,
+                                         -56.2, -59.5, -40.7, -52.7],
+                                        [-57.9, -70.5, -65.5, -65.9, -70.5, -54.9, -66.5, -59.7,
+                                         -54.5, -60.5, -45.1, -60.1],
+                                        [-61.3, -67.8, -65.5, -65.7, -70.4, -55.0, -67.0, -58.5,
+                                         -55.7, -61.4, -52.3, -62.1],
+                                        [-61.8, -67.7, -66.1, -68.1, -70.3, -55.3, -68.4, -58.9,
+                                         -56.5, -64.1, -52.5, -66.7],
+                                        [-62.7, -68.5, -65.9, -67.9, -71.6, -54.9, -68.9, -58.9,
+                                         -55.9, -63.1, -52.7, -67.0],
+                                        [-63.2, -67.6, -64.1, -69.3, -72.7, -54.3, -68.5, -60.5,
+                                         -56.1, -64.5, -54.6, -70.7],
+                                        [-65.0, -66.3, -64.1, -69.7, -72.9, -55.1, -68.3, -60.1,
+                                         -57.3, -64.1, -56.7, -70.9],
+                                        [-67.7, -66.3, -64.0, -69.1, -72.7, -54.7, -67.5, -61.3,
+                                         -60.0, -63.5, -62.0, -71.3],
+                                        [-68.7, -66.4, -63.9, -73.5, -72.5, -56.5, -66.5, -62.6,
+                                         -60.1, -65.7, -62.3, -70.5],
+                                        [-67.3, -67.2, -62.7, -72.3, -67.1, -56.1, -66.0, -63.5,
+                                         -60.0, -65.9, -66.3, -70.9],
+                                        [-66.4, -69.1, -59.1, -73.1, -67.7, -56.6, -62.9, -62.7,
+                                         -59.9, -66.0, -66.3, -71.9],
+                                        [-66.2, -67.5, -59.7, -69.1, -67.7, -59.7, -62.4, -62.3,
+                                         -61.3, -66.2, -71.0, -72.4],
+                                        [-64.5, -64.7, -59.7, -69.0, -65.9, -59.7, -59.7, -59.0,
+                                         -59.9, -66.3, -73.3, -72.7],
+                                        [-68.3, -64.1, -59.9, -68.7, -65.5, -62.1, -59.7, -57.9,
+                                         -59.5, -65.2, -73.9, -70.7],
+                                        [-67.1, -63.3, -58.1, -68.1, -62.1, -62.9, -60.2, -58.7,
+                                         -62.2, -63.1, -74.5, -70.9],
+                                        [-65.1, -62.7, -58.1, -66.1, -61.5, -65.7, -60.7, -59.3,
+                                         -62.7, -62.7, -74.7, -71.9],
+                                        [-65.1, -62.1, -58.1, -64.7, -61.5, -66.3, -60.7, -60.6,
+                                         -63.1, -62.9, -73.1, -70.3],
+                                        [-65.1, -60.7, -58.1, -66.1, -63.3, -66.5, -61.1, -66.5,
+                                         -61.3, -63.1, -76.7, -71.5],
+                                        [-64.7, -60.9, -56.5, -66.1, -63.9, -64.9, -60.4, -69.1,
+                                         -61.9, -64.9, -77.3, -73.3],
+                                        [-63.1, -61.2, -57.5, -65.9, -64.9, -65.9, -59.9, -65.7,
+                                         -61.9, -64.6, -77.1, -68.3],
+                                        [-63.9, -61.3, -56.1, -63.1, -62.9, -65.5, -59.3, -65.5,
+                                         -64.9, -63.3, -75.9, -69.9],
+                                        [-64.3, -60.9, -57.3, -63.0, -62.8, -64.3, -58.4, -65.3,
+                                         -64.7, -63.3, -70.3, -64.5],
+                                        [-66.4, -57.7, -55.7, -62.5, -62.7, -65.0, -57.3, -62.7,
+                                         -63.3, -64.1, -70.9, -63.0],
+                                        [-68.1, -59.3, -53.5, -62.1, -62.7, -66.3, -56.2, -60.8,
+                                         -64.3, -64.0, -64.5, -62.7],
+                                        [-67.4, -58.3, -53.5, -61.8, -60.0, -66.9, -54.9, -60.3,
+                                         -64.3, -62.1, -64.5, -62.1],
+                                        [-62.9, -55.7, -53.7, -61.5, -59.1, -68.5, -55.1, -60.9,
+                                         -59.7, -61.8, -64.5, -62.9],
+                                        [-62.9, -55.7, -53.7, -61.1, -59.5, -67.1, -55.3, -61.5,
+                                         -55.5, -60.9, -62.9, -60.7],
+                                        [-63.9, -55.6, -51.3, -60.5, -59.8, -66.9, -53.9, -61.2,
+                                         -55.1, -61.7, -62.7, -60.3],
+                                        [-63.7, -55.3, -51.5, -59.4, -59.9, -63.1, -52.1, -60.1,
+                                         -55.1, -61.9, -63.1, -60.8],
+                                        [-61.7, -52.5, -52.3, -57.9, -55.7, -62.1, -52.1, -59.7,
+                                         -54.7, -61.6, -62.0, -61.4],
+                                        [-61.1, -51.8, -52.3, -57.9, -53.5, -62.7, -52.5, -56.9,
+                                         -55.2, -60.5, -61.1, -61.9],
+                                        [-61.1, -51.3, -51.5, -57.9, -53.5, -61.9, -51.1, -58.9,
+                                         -56.9, -59.4, -60.9, -59.3],
+                                        [-60.1, -46.9, -50.4, -57.9, -53.5, -60.2, -50.9, -60.3,
+                                         -58.6, -58.2, -60.5, -59.2],
+                                        [-58.3, -46.9, -48.1, -55.3, -53.6, -59.3, -48.3, -60.0,
+                                         -59.7, -57.7, -59.1, -58.9],
+                                        [-60.1, -46.9, -46.9, -54.9, -53.7, -59.3, -48.7, -58.9,
+                                         -58.9, -58.1, -59.8, -58.5],
+                                        [-58.5, -46.9, -46.7, -54.6, -54.0, -60.0, -48.9, -56.3,
+                                         -57.5, -56.3, -60.1, -57.9],
+                                        [-58.5, -46.3, -44.1, -54.3, -54.1, -60.1, -48.3, -57.7,
+                                         -58.7, -56.4, -59.2, -58.3],
+                                        [-57.7, -44.6, -43.4, -53.6, -51.1, -57.4, -46.5, -55.5,
+                                         -58.1, -56.8, -58.1, -59.1],
+                                        [-57.1, -42.9, -42.5, -53.3, -49.3, -57.1, -45.7, -49.3,
+                                         -57.3, -57.2, -58.2, -59.5],
+                                        [-56.4, -41.2, -42.3, -52.9, -48.5, -57.6, -45.7, -49.4,
+                                         -55.1, -57.3, -58.3, -57.5],
+                                        [-53.3, -39.5, -41.3, -52.3, -45.2, -58.7, -45.7, -49.5,
+                                         -54.4, -56.2, -56.5, -57.4],
+                                        [-54.6, -39.7, -39.5, -52.7, -44.1, -57.9, -45.7, -46.5,
+                                         -52.1, -55.5, -55.9, -56.6],
+                                        [-54.7, -40.0, -38.3, -53.7, -45.6, -56.2, -41.1, -46.5,
+                                         -52.3, -55.5, -56.8, -56.5]],
+                               'dims': ('level', 'n2')}},
+ 'dims': {'level': 108, 'n2': 12}}
+
+    ds = xr.Dataset.from_dict(dtemp)
+    return ds
